aws commands: Add region handling to some commands
There were some command on some scripts (I didn't go over all of them
just encountered those during run), that didn't have an explicit region
reference. This works fine only if your default region (which I didn't
set up deliberately and the region you are working on are the same.
This commit simply adds the `--region` parameter explicitly to some
of the aws cli calls.
specifically, you need to have wither `aws` or `region` environment
variable defined in order for the scripts to work.
aws takes precedes over the region variable since we assume that if this
variable exists we are part of a larger context of run that operates on
a specific aws region.

As for the question of why change regions, when using spot instances,
their availability and durability can change depending on the hour and
region :)