add multi-zone scylla-cluster-tests as a submodule
This add a modified version of scylla cluster tests as a submodule.
This will allow to easily create multi zone (or single zone) clusters.
It will also allow us to concentrate the effort in one place instead of
having multiple sources that we need to juggle in order to run the
benchmarks.