ssh1: fix loading phase of workload "a"
By default, and throttling by DynamoDB during the loading phase aborts
it. Change the parameter to retry on throttling (and sleep on the thread
which encountered it). Without this change, the loading phase fails to
load all the data we want it to load.

Also fix an off-by-one error causing one of the loaders to use the wrong
item range and fail.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>