YCSB: add an altered ycsb version as a module
We needed to make some changes to YCSB in order for our
scripts to work. Summary of those changes to date:
1. Use the credentials n the default profile of the host - in order to
   use temporary credentials
2. Bump the HdrHistograms version in order to be able to work with them
   in python.
3. Add a missing library in order to be able to run ycsb on Java later
   than JAVA 9.0

Here we add this version of YCSB as a submodule.

Also added here a script `make_ycsb` that will build the dynamodb and
scylla binding and extract them to a folder ./ycsb_exec in order to get
uniformity and reproducibility in our ycsb runs.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>