alternator-benchmark
This script is based on the script ssh1 from this repository,
It comes bundled with additional script for results processing
results and it's python requirements.

The main differences from ssh1 are (not exhaustive):
1. It has a bunch of parameters to control it's behavior,
./alternator-benchmark -h/--help for the full list and explanations.

2. It saves the results on persistent storage along with a lot of
information about the run and the stats in a way that will make
reproducing or retrospecting the runs easy.

3. It processes the results into readable text files.

4. It has the ability to run on scylla too.

The script is probably rough around the edges and will require
additional work but I have tested it on some workloads.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>