Introduce a script to disable snapshots
When doing benchmarks of Scylla and Alternator there is no real value in
snapshots (for DROP TABLE), it can however prevent rapid bencmarking
where you would like to first delete the tables before loading them.
This commit adds a script to disable the snapshoting in a Scylla cluster
and also calls it in the `create_cluster` script, which means that
cluster created with this script will have snapshots disabled by
default.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>